requirements.txt
Django==4.1
django-filter==22.1
django-phonenumber-field==6.3.0
djangorestframework==3.13.1
celery==5.2.7
requests==2.25.1
django-celery-results==2.4.0
redis==4.3.4
django-redis==5.2.0



